added items navigation
