+++ app/src/main/java/com/zeljkostankovic/dalitalk/fragments/items/ItemsAddFragment.java
@@ -1,4 +1,4 @@
-package com.zeljkostankovic.dalitalk.fragments;
+package com.zeljkostankovic.dalitalk.fragments.items;
 
 import android.os.Bundle;
 
@@ -12,10 +12,10 @@ import com.zeljkostankovic.dalitalk.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link HomeFragment#newInstance} factory method to
+ * Use the {@link ItemsAddFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class HomeFragment extends Fragment {
+public class ItemsAddFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -26,7 +26,7 @@ public class HomeFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public HomeFragment() {
+    public ItemsAddFragment() {
         // Required empty public constructor
     }
 
@@ -36,11 +36,11 @@ public class HomeFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment HomeFragment.
+     * @return A new instance of fragment ItemsAddFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static HomeFragment newInstance(String param1, String param2) {
-        HomeFragment fragment = new HomeFragment();
+    public static ItemsAddFragment newInstance(String param1, String param2) {
+        ItemsAddFragment fragment = new ItemsAddFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -61,6 +61,6 @@ public class HomeFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false);
+        return inflater.inflate(R.layout.fragment_items_add, container, false);
     }
 }
