+++ app/src/main/java/com/zeljkostankovic/dalitalk/fragments/items/ItemsListFragment.java
@@ -1,0 +1,34 @@
+package com.zeljkostankovic.dalitalk.fragments.items;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+import androidx.navigation.Navigation;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.zeljkostankovic.dalitalk.R;
+
+public class ItemsListFragment extends Fragment {
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        View view = inflater.inflate(R.layout.fragment_items_list, container, false);
+
+        FloatingActionButton fab = (FloatingActionButton) view.findViewById(R.id.fabItemsList);
+        fab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Navigation.findNavController(v).navigate(R.id.action_itemsListFragment_to_itemsAddFragment);
+            }
+        });
+
+        return view;
+    }
+}
