+++ app/src/main/res/layout/fragment_items_add.xml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/fragmentHome"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/cardview_dark_background"
-    tools:context=".fragments.home.HomeFragment">
+    tools:context=".fragments.items.ItemsAddFragment">
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
