+++ app/src/main/res/navigation/navigation.xml
@@ -7,12 +7,34 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.zeljkostankovic.dalitalk.fragments.HomeFragment"
+        android:name="com.zeljkostankovic.dalitalk.fragments.home.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/configFragment"
-        android:name="com.zeljkostankovic.dalitalk.fragments.ConfigFragment"
+        android:name="com.zeljkostankovic.dalitalk.fragments.config.ConfigFragment"
         android:label="fragment_config"
-        tools:layout="@layout/fragment_config" />
+        tools:layout="@layout/fragment_config" >
+        <action
+            android:id="@+id/action_configFragment_to_itemsListFragment"
+            app:destination="@id/itemsListFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/itemsListFragment"
+        android:name="com.zeljkostankovic.dalitalk.fragments.items.ItemsListFragment"
+        android:label="fragment_items_list"
+        tools:layout="@layout/fragment_items_list" >
+        <action
+            android:id="@+id/action_itemsListFragment_to_itemsAddFragment"
+            app:destination="@id/itemsAddFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/itemsAddFragment"
+        android:name="com.zeljkostankovic.dalitalk.fragments.items.ItemsAddFragment"
+        android:label="fragment_items_add"
+        tools:layout="@layout/fragment_items_add" >
+        <action
+            android:id="@+id/action_itemsAddFragment_to_itemsListFragment"
+            app:destination="@id/itemsListFragment" />
+    </fragment>
 </navigation>
